+++ app/Controllers/BlogController.php
@@ -14,6 +14,6 @@ class BlogController extends Base
     public function dispatch()
     {
         $this->page->setTitle(__('page.blog'));
-        $this->page->disableJavascriptCheck();
+        $this->jsCheck = false;
     }
 }