+++ app/Controllers/CommunityController.php
@@ -10,6 +10,7 @@ class CommunityController extends Base
     public function dispatch()
     {
         $this->page->setTitle(__('page.explore'));
+        $this->jsCheck = false;
 
         if (isLogged() && $this->fetchGet('i')) {
             $post = Post::where('server', $this->fetchGet('s'))