+++ app/Controllers/NotfoundController.php
@@ -9,6 +9,6 @@ class NotfoundController extends Base
     public function dispatch()
     {
         $this->page->setTitle(__('page.not_found'));
-        $this->page->disableJavascriptCheck();
+        $this->jsCheck = false;
     }
 }